feat: edge application error responses progress
